+++ saas_demo/doc/changelog.rst
@@ -2,6 +2,7 @@
 -------
 
 - **Fix:** errors on manual creation of ``saas.demo`` record
+- **Fix:** git pull doesn't work when odoo is not a git folder
 
 `1.0.0`
 -------